refactor(apex-fixes): rename src/ → lib/ to match plugin convention
