+++ plugins/apex-fixes/lib/privileged.ts
