+++ plugins/apex-fixes/lib/sleep-enable.ts
