+++ plugins/apex-fixes/lib/xhci-recovery.ts
